Participate in chat protocol only through chat CLI
